Add LIMITS_NO_ALARM_ON_INACTIVE configuration option.
By default, any toggle on a hard limit pin will cause a hard
limit alarm. This option will change this to not cause an
alarm when a pin goes from active-to-inactive.
It should be safe to enable this option by default, even if
you have bouncing switches, as the switch will reach a stable
state eventually.