Add ABORT_LOCK configuration option.
Useful when ESTOP is connected to Grbl's Abort/Reset pin.
Will force an alarm when ESTOP is activated, and prevents
exiting the alarm mode (either through $H or $X) as long
as it stays active.
Note: enabling this option will always force an alarm when
the pin is activated (i.e. even when there is no active
motion), as it is assumed that power to your drivers is also
disabled on ESTOP, so e.g. microsteps are lost.
Enabling this option is transparent when the abort pin is
left unconnected (unless INVERT_CONTROL_PIN is used).